Feat/상품서비스 서비스레이어구현

+++ src/main/java/com/yeonieum/productservice/domain/category/repository/ProductDetailCategoryRepository.java
@@ -3,5 +3,5 @@ package com.yeonieum.productservice.domain.category.repository;
 import com.yeonieum.productservice.domain.category.entity.ProductDetailCategory;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductDetailCategoryRepository extends JpaRepository<ProductDetailCategory,Integer> {
+public interface ProductDetailCategoryRepository extends JpaRepository<ProductDetailCategory,Long> {
 }
